+++ sReports/sReportsV2.Domain/Entities/Field/FieldNumeric.cs
@@ -1,10 +1,5 @@
 ﻿using MongoDB.Bson.Serialization.Attributes;
 using sReportsV2.Common.Constants;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace sReportsV2.Domain.Entities.FieldEntity
 {
@@ -13,8 +8,8 @@ namespace sReportsV2.Domain.Entities.FieldEntity
     public class FieldNumeric : FieldString
     {
         public override string Type { get; set; } = FieldTypes.Number;
-        public int? Min { get; set; }
-        public int? Max { get; set; }
+        public double? Min { get; set; }
+        public double? Max { get; set; }
         public double? Step { get; set; }
     }
 }
